requirements.txt
requests >= 2.20.0
Pillow >= 7.0.0